Multiple fixes. See description.
- Direct light fix: previously we returned a vector of ray scalars for intersections. The method that checks if there was or not a shadow ray was wrong, it checked that if t > 0.01f and t < 1 then it collisioned with that object, thats incorrect, the SHADOW BIAS is used to correct the intersection point with the normal direction. Then you have only to check if t > 0. Also, the direct light needed the kd/PI that was ommited in previous versions.

- Indirect light: the russian roulette coeficient doesn't multiply (direct + indirect), just indirect.

- Sphere collision: I don't know why, but no version was correct for our implemention... I found one that finally works but to work I needed to change float to double.

- Refraction fix: due to rounding errors, this doesn't worked, also that the methods that I founded worked bad... This is already fixed.

- Fresnel coefficient: work in progress.

- Area lights: next station.

- Need to fix previous practices, that because I've change the core libraries :)

+++ practica1/main.cpp
@@ -4,7 +4,7 @@
 void geometry_test(bool verbose = false) {
     // Vector3 tests.
     std::cout << "== PRUEBA DE VECTOR3 =================" << std::endl;
-    float s = 0.5;
+    double s = 0.5;
     Vector3 a(1.0,1.0,1.0), b(5.0,4.0,4.5), d(2.0, 2.0, 2.0), v(2.0, -2.0, 0.0), w = crs(d,v);
 
     if (verbose) { std::cout << "Params: " << a << ", " << b << ", " << s << std::endl; }
@@ -21,8 +21,8 @@ void geometry_test(bool verbose = false) {
 
     // Matrices tests.
     std::cout << "== PRUEBA DE MATRICES =======================" << std::endl;
-    float A[4][4] = {{3,4,7,8},{3,2,8,7},{0,2,6,7},{4,2,2,1}};
-    float B[4][4] = {{8,7,6,5},{2,4,2,1},{1,1,3,4},{7,8,9,0}};
+    double A[4][4] = {{3,4,7,8},{3,2,8,7},{0,2,6,7},{4,2,2,1}};
+    double B[4][4] = {{8,7,6,5},{2,4,2,1},{1,1,3,4},{7,8,9,0}};
 
     Matrix3 mA(A);
     Matrix3 mB(B);
@@ -100,7 +100,7 @@ int main (int argc, char* argv[]) {
     // Si hay sugerencia de alternativas de nombre de Matrix3, soy todo oidos.
     Planet A(Vector3(1,1,1), Vector3(0,0,0), Vector3(2,2,2)), 
         B(Vector3(1,1,1), Vector3(5,5,5), Vector3(7,7,7));
-    float latA, aziA, latB, aziB;
+    double latA, aziA, latB, aziB;
 
     // PLANET A:
     std::cout << "Welcome to the FTL Enterprises Interplanetary Simulator" << std::endl;
